+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Moved to use GitHub Actions for label enforcement
+- Add extra flags to Intel Fortran Debug builds to make them stricter
+  - `-stand f18` to enable Fortran 2018 Standard compliance
+  - `-diag-error 6188` to cause if(integer) to fail
+  - `-diag-error 6192` to cause logical set to integer to fail
+  - `-diag-disable 5268` to suppress warning for long source lines (which our macros often make)
 
 ### Fixed